lab4_cloudnet/requirements.txt
Django
django-webpack-loader
djangorestframework
mysqlclient
django-filter
django-livereload
python-memcached
uWSGI
django-elasticache
